refactor: extract computation logic to custom hook
